backend/requirements.txt
absl-py==0.11.0
appdirs==1.4.4
astunparse==1.6.3
black==20.8b1
bleach==3.2.1
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.1.1
cycler==0.10.0
defusedxml==0.6.0
Django==2.2.16
django-allauth==0.43.0
django-appconf==1.0.4
django-cors-headers==3.5.0
django-resized==0.3.11
django-rest-auth==0.9.5
django-rest-authtoken==2.1.3
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
docutils==0.16
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==21.4.0
kiwisolver==1.3.1
Markdown==3.3.3
matplot==0.1.9
matplotlib==3.3.2
mypy-extensions==0.4.3
mysqlclient==2.0.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.4
pathspec==0.8.0
Pillow==8.0.1
pkginfo==1.6.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
pyloco==0.0.139
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
pywin32-ctypes==0.2.0
PyYAML==5.3.1
readme-renderer==28.0
regex==2020.10.23
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.6
scipy==1.5.3
SimpleWebSocketServer==0.1.1
six==1.15.0
sqlparse==0.4.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
toml==0.10.1
tqdm==4.51.0
twine==3.2.0
typed-ast==1.4.1
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.25.11
ushlex==0.99.1
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
